requirements-prod.txt
Flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
requests==2.31.0
PyYAML==6.0.1
Werkzeug==3.0.1
PyPDF2==3.0.1
pandas==2.1.4
openpyxl==3.1.2